refactor: #305 linkuService impl 다 합침
