+++ src/main/java/com/umc/linkyou/service/Linku/LinkuSearchService.java
@@ -1,8 +1,24 @@
 package com.umc.linkyou.service.Linku;
 
+import com.umc.linkyou.repository.linkuRepository.LinkuRepository;
+import com.umc.linkyou.repository.linkuRepository.LinkuRepositoryCustom;
 import com.umc.linkyou.web.dto.linku.LinkuSearchSuggestionResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 
-public interface LinkuSearchService {
-    List<LinkuSearchSuggestionResponse> suggest(Long userId, String keyword);
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class LinkuSearchService{
+
+    private final LinkuRepository linkuRepository;
+
+    public List<LinkuSearchSuggestionResponse> suggest(Long userId, String keyword) {
+        String q = (keyword == null) ? "" : keyword.trim();
+        if (q.length() < 2) return List.of(); // 입력 2자 미만 가드
+        return linkuRepository.findUserSavedSuggestions(userId, q);
+    }
 }
