--- src/main/java/com/umc/linkyou/service/Linku/LinkuSearchServiceImpl.java
@@ -1,25 +1,0 @@
-package com.umc.linkyou.service.Linku;
-
-import com.umc.linkyou.repository.linkuRepository.LinkuRepository;
-import com.umc.linkyou.repository.linkuRepository.LinkuRepositoryCustom;
-import com.umc.linkyou.web.dto.linku.LinkuSearchSuggestionResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class LinkuSearchServiceImpl implements LinkuSearchService {
-
-    private final LinkuRepository linkuRepository;
-
-    @Override
-    public List<LinkuSearchSuggestionResponse> suggest(Long userId, String keyword) {
-        String q = (keyword == null) ? "" : keyword.trim();
-        if (q.length() < 2) return List.of(); // 입력 2자 미만 가드
-        return linkuRepository.findUserSavedSuggestions(userId, q);
-    }
-}
